changed day 8 values from long to int (it doesn't affect outcome, but it fits the assignment better)

+++ src/main/java/net/kenvanhoeylandt/solutions/day7/logic/Gate.java
@@ -24,5 +24,5 @@ public abstract class Gate
 		return mGateManager;
 	}
 
-	abstract public long getValue();
+	abstract public int getValue();
 }
