+++ src/main/java/net/kenvanhoeylandt/solutions/day7/logic/ValueProvider.java
@@ -2,5 +2,5 @@ package net.kenvanhoeylandt.solutions.day7.logic;
 
 public interface ValueProvider
 {
-	long getValue();
+	int getValue();
 }
