+++ src/main/java/net/kenvanhoeylandt/solutions/day7/logic/gates/RightShiftGate.java
@@ -14,7 +14,7 @@ public class RightShiftGate extends SingleInputGate
 	}
 
 	@Override
-	public long getValue()
+	public int getValue()
 	{
 		return getInputValue() >> mShift;
 	}
